requirements.txt
discord.py
mysql
mysql-connector